refactor: LauncherCompanion を LauncherAgent にリネーム (#30 #214)
Companions/ 内の命名不整合を解消。`Companions/LauncherCompanion/` は (a) カテゴリ名
"Companions" を member 名で繰り返すスタッター、(b) `Updater` (機能名) と異なり consumer 軸 +
意味の薄い "Companion" 命名、で浮いていた。"Agent" (常駐ヘルパーの定番語) に変えてスタッター
解消 + 読み心地を Updater と揃える。

判断軸の明確化: Companion 系を分割/統合するかは「責務の数」でなく「ライフサイクル/結合度」で
決める。Updater が別 exe なのはハード制約 (実行中 exe は自分を置換できない + 親が落ちている間に
動く)。LauncherAgent の probe/sensor/focus は Launcher と同ライフサイクル + 単一 UDP + 単一
イベントループを共有するので統合が正しい。

- Companions/LauncherCompanion/ → Companions/LauncherAgent/ (folder/csproj/AssemblyName/
  namespace TonePrism.LauncherAgent/exe TonePrism_LauncherAgent.exe、git mv で履歴保持)
- Godot: autoload LauncherCompanion → LauncherAgent、scripts/launcher_companion.gd →
  launcher_agent.gd、呼出箇所 (game_session/overlay_manager/game_selection) と _EXE_CANDIDATES
  パス、ログ prefix [LauncherAgent] を更新
- Release.ps1: Build-LauncherAgent / staging path / Assert-ExpectedFiles を更新
- Manager: 参照コメント (UpdateSectionPanel/UpdateDownloader/LogLineFormat) を更新
- CHANGELOG 既存履歴は当時の事実として残置 (Stage 5 でリネーム注記を追加予定)
