+++ Companions/LauncherAgent/App.config
